Changes Entity containers to contain pointers to entities, instead of actual entities.
This was done to enable the creation of arrays of mutex inside food
class. As mutex is not a movable or copyable data type, we cant
push_back foods in m_foods at their creation. This motivated the
change for m_foods, which, in turn, motivated the change to the rest
of the containers.

This commit also includes initial implementation of the philosophers
dinner problem inside food sources.